refactor: move privacy policy folder to be in legal folder
