+++ src/test/legal/privacyPage/PrivacyPage.test.tsx
